# topgun-services/docs/swagger.yaml
definitions:
  models.Camera:
    properties:
      id:
        type: string
      institute:
        type: string
      location:
        type: string
      name:
        type: string
      token:
        type: string
    type: object
...
  models.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  models.ResetPasswordRequest:
    properties:
      new_password:
        minLength: 6
        type: string
      old_password:
        type: string
    required:
    - new_password
    - old_password
    type: object
  models.User:
    properties:
      email:
        type: string
      first_name:
        type: string
      first_name_en:
        type: string
      last_name:
        type: string
      last_name_en:
        type: string
      password:
        type: string
      title:
        type: string
      title_en:
        type: string
    type: object
  models.VerifyTokenAndResetPasswordRequest:
    properties:
      password:
        type: string
      token:
        type: string
    type: object
info:
  contact: {}
  title: KKU GS ADMISSION SERVICES API
paths:
  /api/v1/auth/login:
    post:
      consumes:
      - application/json
      description: Login a user
      parameters:
      - description: Login Request
        in: body
        name: login
        required: true
        schema:
          $ref: '#/definitions/models.LoginRequest'
      produces:
      - application/json
      responses: {}
      summary: Login
      tags:
      - Auth
  /api/v1/auth/register:
    post:
      consumes:
      - application/json
      description: Register a new user
      parameters:
      - description: User
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.User'
      produces:
      - application/json
      responses: {}
      summary: Register
      tags:
      - Auth
  /api/v1/auth/reset-password:
    post:
      consumes:
      - application/json
      description: Reset user password with old and new password
      parameters:
      - description: Reset Password Request
        in: body
        name: resetPassword
        required: true
        schema:
          $ref: '#/definitions/models.ResetPasswordRequest'
      produces:
      - application/json
      responses: {}
      security:
      - ApiKeyAuth: []
      summary: Reset Password
      tags:
      - Auth
  /api/v1/auth/reset-password/{email}/issue-token:
    post:
      consumes:
      - application/json
      description: Issue token for email verification
      parameters:
      - description: Email
        in: path
        name: email
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      summary: Issue token for email verification
      tags:
      - Auth
  /api/v1/auth/reset-password/verify:
    post:
      consumes:
      - application/json
      description: Verify token and reset password
      parameters:
      - description: VerifyTokenAndResetPasswordRequest
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.VerifyTokenAndResetPasswordRequest'
      produces:
      - application/json
      responses: {}
      summary: Verify token and reset password
      tags:
      - Auth
  /api/v1/camera/:
    get:
      consumes:
      - application/json
      description: Get list of cameras with pagination and search
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Items per page
        in: query
        name: limit
        type: integer
      - description: Search term
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses: {}
      security:
      - ApiKeyAuth: []
      summary: GetCameras
      tags:
      - Camera
    post:
      consumes:
      - application/json
      description: Create a new camera
      parameters:
      - description: Camera object
        in: body
        name: camera
        required: true
        schema:
          $ref: '#/definitions/models.Camera'
      produces:
      - application/json
      responses: {}
      security:
      - ApiKeyAuth: []
      summary: CreateCamera
      tags:
      - Camera
  /api/v1/camera/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a camera by ID
      parameters:
      - description: Camera ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      security:
      - ApiKeyAuth: []
      summary: DeleteCamera
      tags:
      - Camera
    get:
      consumes:
      - application/json
      description: Get a camera by ID
      parameters:
      - description: Camera ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      security:
      - ApiKeyAuth: []
      summary: GetCamera
      tags:
      - Camera
    put:
      consumes:
      - application/json
      description: Update an existing camera
      parameters:
      - description: Camera ID
        in: path
        name: id
        required: true
        type: string
      - description: Camera object
        in: body
        name: camera
        required: true
        schema:
          $ref: '#/definitions/models.Camera'
      produces:
      - application/json
      responses: {}
      security:
      - ApiKeyAuth: []
      summary: UpdateCamera
      tags:
      - Camera
...
schemes:
- http
- https
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"